# docs/external/prdb-openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "prdb.PublicApi | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5080/"
    }
  ],
  "paths": {
    "/wanted-videos/{videoId}": {
      "put": {
        "tags": [
          "UpdateWantedVideo"
        ],
        "summary": "Update a wanted video",
        "description": "Updates the fulfilment fields of a wanted video entry for the currently authenticated user. Set IsFulfilled to true to mark as fulfilled, or false to reset it.",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWantedVideoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWantedVideoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateWantedVideoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RemoveWantedVideo"
        ],
        "summary": "Remove a wanted video",
        "description": "Removes the specified video from the wanted list of the currently authenticated user. Returns 404 if the video was not on the user's wanted list. Requires API key authentication.",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": { }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AddWantedVideo"
        ],
        "summary": "Add a wanted video",
        "description": "Adds the specified video to the wanted list for the currently authenticated user. Returns 204 if added or already on the wanted list. Returns 404 if the video does not exist. Requires API key authentication.",
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": { }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/wanted-videos": {
      "get": {
        "tags": [
          "ListWantedVideos"
        ],
        "summary": "List wanted videos",
        "description": "Returns a paged, sortable list of wanted videos for the currently authenticated user. Supports filtering by fulfilment status, site, actor, search term, and recency. Requires API key authentication.",
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 500,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "pattern": "^(?i)(asc|desc)$",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "description": "Allowed values: asc, desc."
            }
          },
          {
            "name": "IsFulfilled",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SiteId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ActorId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AddedInLastDays",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListWantedVideosResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/videos": {
      "get": {
        "tags": [
          "ListVideos"
        ],
        "summary": "List videos",
        "description": "Returns a paged list of videos. Supports filtering by site ID and searching by video title or site title. Requires API key authentication.",
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SiteId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ActorId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "pattern": "^(?i)(asc|desc)$",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "description": "Allowed values: asc, desc."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListVideosResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/videos/{id}": {
      "get": {
        "tags": [
          "GetVideoById"
        ],
        "summary": "Get video by ID",
        "description": "Returns full details for a single video, including site, network, images, alternative titles, and actors with their images. Requires API key authentication.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActorGender": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "description": "Known values: Unknown (0), Female (1), Male (2), Intersex (3), Transmale (4), Transfemale (5), NonBinary (6)."
      },
...
      "ActorImageType": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "description": "Known values: Thumbnail (1), Poster (2), Face (3)."
      },
...
      "FulfillmentApp": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "description": "Known values: Sabnzbd (0), Nzbget (1), Filesystem (2), Other (3)."
      },
...
      "ListVideosResponse": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize",
          "sortBy",
          "sortDirection"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoSummaryDto"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "sortBy": {
            "type": "string"
          },
          "sortDirection": {
            "type": "string"
          }
        }
      },
      "ListWantedVideosResponse": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize",
          "totalPages",
          "sortBy",
          "sortDirection"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WantedVideoSummaryDto"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "sortBy": {
            "type": "string"
          },
          "sortDirection": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateWantedVideoRequest": {
        "type": "object",
        "properties": {
          "isFulfilled": {
            "type": "boolean"
          },
          "fulfilledAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "fulfilledInQuality": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VideoQuality"
              }
            ]
          },
          "fulfillmentExternalId": {
            "type": [
              "null",
              "string"
            ]
          },
          "fulfillmentByApp": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FulfillmentApp"
              }
            ]
          }
        }
      },
      "UpdateWantedVideoResponse": {
        "required": [
          "videoId",
          "isFulfilled",
          "createdAtUtc",
          "updatedAtUtc"
        ],
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string",
            "format": "uuid"
          },
          "isFulfilled": {
            "type": "boolean"
          },
          "fulfilledAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "fulfilledInQuality": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VideoQuality"
              }
            ]
          },
          "fulfillmentExternalId": {
            "type": [
              "null",
              "string"
            ]
          },
          "fulfillmentByApp": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FulfillmentApp"
              }
            ]
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "VideoDetailActorDto": {
        "required": [
          "id",
          "name",
          "gender",
          "nationality",
          "images"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "gender": {
            "$ref": "#/components/schemas/ActorGender"
          },
          "birthday": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "nationality": {
            "$ref": "#/components/schemas/Nationality"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoDetailActorImageDto"
            }
          }
        }
      },
      "VideoDetailActorImageDto": {
        "required": [
          "id",
          "imageType"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cdnPath": {
            "type": [
              "null",
              "string"
            ]
          },
          "imageType": {
            "$ref": "#/components/schemas/ActorImageType"
          }
        }
      },
      "VideoDetailDto": {
        "required": [
          "id",
          "title",
          "createdAtUtc",
          "updatedAtUtc",
          "site",
          "images",
          "preNames",
          "actors"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "releaseDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "site": {
            "$ref": "#/components/schemas/VideoDetailSiteDto"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoDetailImageDto"
            }
          },
          "preNames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoDetailPreNameDto"
            }
          },
          "actors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoDetailActorDto"
            }
          }
        }
      },
      "VideoDetailImageDto": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cdnPath": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "VideoDetailNetworkDto": {
        "required": [
          "id",
          "title",
          "url"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "VideoDetailPreNameDto": {
        "required": [
          "id",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "VideoDetailSiteDto": {
        "required": [
          "id",
          "title",
          "url"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "network": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VideoDetailNetworkDto"
              }
            ]
          }
        }
      },
...
      "VideoQuality": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "description": "Known values: P720 (0), P1080 (1), P2160 (2)."
      },
      "VideoSummaryActorDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "VideoSummaryDto": {
        "required": [
          "id",
          "title",
          "siteId",
          "siteTitle",
          "actors"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "siteId": {
            "type": "string",
            "format": "uuid"
          },
          "siteTitle": {
            "type": "string"
          },
          "releaseDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "actors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoSummaryActorDto"
            }
          }
        }
      },
...
      "WantedVideoSummaryDto": {
        "required": [
          "videoId",
          "videoTitle",
          "siteTitle",
          "isFulfilled",
          "createdAtUtc",
          "updatedAtUtc"
        ],
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string",
            "format": "uuid"
          },
          "videoTitle": {
            "type": "string"
          },
          "siteTitle": {
            "type": "string"
          },
          "videoReleaseDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "videoCreatedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "imageCdnPath": {
            "type": [
              "null",
              "string"
            ]
          },
          "isFulfilled": {
            "type": "boolean"
          },
          "fulfilledAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "fulfilledInQuality": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VideoQuality"
              }
            ]
          },
          "fulfillmentExternalId": {
            "type": [
              "null",
              "string"
            ]
          },
          "fulfillmentByApp": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FulfillmentApp"
              }
            ]
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
...
}